lesson 3

+++ server/src/routes/memories.ts
@@ -1,10 +1,17 @@
 import { FastifyInstance } from 'fastify';
 import { z } from 'zod';
-import { prisma } from '../lib/prisma';
+import { prisma } from './../lib/prisma';
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify();
+    });
+
+    app.get('/memories', async (request) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                id: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             },
@@ -19,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         });
     });
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, response) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -29,6 +36,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             },
         });
+
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return response.status(401).send();
+        }
 
         return memory;
     });
@@ -47,14 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '023faedb-e63f-4ba3-ab54-d04972424f07',
+                userId: request.user.sub,
             },
         });
 
         return memory;
     });
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, response) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -65,7 +76,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
             isPublic: z.coerce.boolean().default(false),
         });
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            },
+        });
+        if (memory.userId !== request.user.sub) {
+            return response.status(401).send();
+        }
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -79,13 +98,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory;
     });
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, response) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
-
         const { id } = paramsSchema.parse(request.params);
-
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            },
+        });
+        if (memory.userId !== request.user.sub) {
+            return response.status(401).send();
+        }
         await prisma.memory.delete({
             where: {
                 id,
